+++ src/configuration/languageIds.ts
@@ -16,6 +16,7 @@ export const typescriptreact = 'typescriptreact';
 export const javascript = 'javascript';
 export const javascriptreact = 'javascriptreact';
 export const jsxTags = 'jsx-tags';
+export const assemblyscript = 'assemblyscript';
 
 export const jsTsLanguageModes = [
     javascript,